pyproject.toml
[project]
name = "selfheal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "hypothesis>=6.152.1",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.2",
    "langgraph>=1.1.9",
    "pypdf>=6.10.2",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "uvicorn>=0.46.0",
]
